Added GlutWindow* classes. Minor fixes to make code equal to used in "Computer Graphics: From Pixels to Programmable Graphics Hardware"

+++ include/DdsDefs.h
@@ -5,34 +5,35 @@
 #ifndef	__DDS_TYPES__
 #define	__DDS_TYPES__
 
+#include	"TypeDefs.h"
 #pragma pack (push, 1)
 
 struct DDS_PIXELFORMAT
 {
-        unsigned long dwSize;
-        unsigned long dwFlags;
-        unsigned long dwFourCC;
-        unsigned long dwRGBBitCount;
-        unsigned long dwRBitMask;
-        unsigned long dwGBitMask;
-        unsigned long dwBBitMask;
-        unsigned long dwABitMask;
+        uint32 dwSize;
+        uint32 dwFlags;
+        uint32 dwFourCC;
+        uint32 dwRGBBitCount;
+        uint32 dwRBitMask;
+        uint32 dwGBitMask;
+        uint32 dwBBitMask;
+        uint32 dwABitMask;
 };
 
 struct DDS_HEADER
 {
-        unsigned long dwSize;
-        unsigned long dwFlags;
-        unsigned long dwHeight;
-        unsigned long dwWidth;
-        unsigned long dwPitchOrLinearSize;
-        unsigned long dwDepth;
-        unsigned long dwMipMapCount;
-        unsigned long dwReserved1[11];
+        uint32 		dwSize;
+        uint32 		dwFlags;
+        uint32 		dwHeight;
+        uint32 		dwWidth;
+        uint32 		dwPitchOrLinearSize;
+        uint32 		dwDepth;
+        uint32 		dwMipMapCount;
+        uint32 		dwReserved1[11];
         DDS_PIXELFORMAT ddspf;
-        unsigned long dwCaps1;
-        unsigned long dwCaps2;
-        unsigned long dwReserved2[3];
+        uint32 		dwCaps1;
+        uint32 		dwCaps2;
+        uint32		dwReserved2[3];
 };
 
 #pragma	pack (pop)
