+++ include/TypeDefs.h
@@ -17,13 +17,12 @@
 	#include <stdint.h>
 #endif
 
-		// XXX
 typedef	unsigned char		uint8;
 typedef	signed	 char		int8;
 typedef	unsigned short		uint16;
 typedef	signed	 short		int16;
-typedef	unsigned long		uint32;
-typedef	signed	 long		int32;
+typedef	unsigned int		uint32;
+typedef	signed	 int		int32;
 
 #ifdef	_WIN32
 typedef	unsigned __int64	uint64;
@@ -33,7 +32,7 @@ typedef	unsigned long long	uint64;
 typedef signed   long long	int64;
 #endif
 
-typedef	unsigned char		byte;
+typedef	unsigned char	byte;
 typedef	uint16			word;
 typedef	uint32			dword;
